♻️ [Refactor]: material, chat component 디렉토리 변경
app -> entity
